requirements.txt
google-ads>=25.0.0
pytrends>=4.8.0
pandas>=2.2.0
numpy>=1.26.0
matplotlib>=3.8.0
plotly>=5.20.0
jinja2>=3.1.0
weasyprint>=61.0
pydantic>=1.10.0
python-dotenv>=1.0.0
tqdm>=4.66.0
typer[all]>=0.12.0
pytest>=8.0.0
requests>=2.31.0
